critical: do not OVERWRITE a relay's (existing) keys
In the last commit offlinemasterkey functionality has been introduced.
Keys are generated on the ansible host and copied to the relay, but if
the relay had already keys in place these keys would have been overwritten.

This commit simply disables that task, we will come up with a check to
see whether the relay has keys already that should be copied to the ansible host
before proceeding so that the relay's keys are not overwritten and the relay
identity does not change.

DO NOT use this branch for anything real yet!