add safetycheck: detect and abort on RSA key mismatch (remote vs. local)
Since we do not override remote RSA keys we might end up in a
situation where the local key does not match the remote key.
That would result in all kind of troubles (MyFamily fingerprints become incorrect).

If we detect such a case we abort and require manual intervention.

The user has two options:
- Trust the remote key and replace the local key.
- Replace the remote key with the local key.

It is up to the user to decide and resolve manually.