major workaround: make more than two instances per host actually work (#153)
Due to a bug in tor, instances that are not binding on the first
public IPv4 address are not announcing their correct IP address
(they use the first IP instead of the one configured via `OutboundBindAddress`)
and therefor fail their reachability test and never upload a valid descriptor.

This commit works around this by explicitly setting `Address` to the same
IP address as the one used for `OutboundBindAddress`.

Once the upstream bugs are fixed we should revert this change so
we might also work in setups where public IPs are NATed to other public IPs.

This bug affected all users which set `tor_maxPublicIPs` to a value greater than 1.

Upstream bugs:
https://trac.torproject.org/projects/tor/ticket/13953
https://trac.torproject.org/projects/tor/ticket/19919

Thanks to Tim Niemeyer (f3netze.de) for bringing this to our attention.