support multiple IP addresses for use with OutboundBindAddressExit
before this commit we used a single IP address for all tor instances
on a server for exiting if tor_dedicatedExitIP was enabled, with this
commit multiple exit IP addresses are used. By distributing the
traffic across multiple source IP addresses this should help reduce
the negative effects of bad traffic spoiling the source IP for many
tor users.

There are two cases to consider:
- user runs one tor instance per IP (tor_ports has one entry):
"loop_idx" can be directly used as an index to the tor_available_public_ipv4/6 arrays
- user runs two tor instances per IP (tor_ports has two entries):
in this case we need to mangle "loop_idx" before we used it as an index
to the IP address arrays

This change includes also the kitchen test suites for this feature:
t-exit-6publicIPs-dedicatedExitIP-2instPerIP
t-exit-6publicIPs-dedicatedExitIP-singleInstPerIP