change the ControlSocket path
Before this change the ControlSocket file resided in the instance's
DataDir. At the same time that directory had permission 700
so it was basically unusable unless you did access it as root (bad)
or the user running tor itself (also bad).

Now the ControlSocket file is located in the
/var/run/tor-instances subfolders
to make it actually accessible to non-root users.