MetricsPort: generate htpasswd file on relay
previously we generated the htpasswd file on the
ansible control machine and left it to the user
to copy it to the relay.

it can be enabled via tor_gen_metricsport_htpasswd

When enable we also install the dependencies
(python passlib) on the server.

The file will be written to the path defined in
tor_metricsport_htpasswd_file
and owned by root:www-data (debian)
or root:www (BSD)